requirements.txt
pymodbus==3.7.0
python-dotenv==1.0.0
pyserial==3.5